requirements.txt
python-telegram-bot==12.7
pandas==1.3.0
psycopg2==2.9.1